+++ man-roxygen/gapParam.R
@@ -1,0 +1,2 @@
+#' @param gap An unambiguous abbreviation of `inapplicable`, `ambiguous` 
+#' (= `missing`), or `extra state`, specifying how gaps will be handled.